Add full support for multi-unit SUT for AAA
Alternate activity allocation rests on the assumption that the
secondary production of a commodity can be approximated based on the
recipe of a primary production.  This poses a problem for multi-unit
SUT. One technology is assumed to be equivalent for producing two
products... in what units?  Update adds possibility of 3D Gamma
variable, with the middle dimension equal to the set of properties that
describe the products, and an optional argument for these properties

Allows for alternate technology assumptions when products are described
simultaneously in terms of multiple properites. Makes the Alternate
Activity matrix 3-dimensional:

	gamma_{I,m,i}:
		 the requirements of industry I, normalized in terms of
		 property m,  will be assumed for each secondary unit of
		 production of product i, also measured in terms of m.

	For example: the requirements of cattle production (dimension 1)
	per kg of its output (mass: dimension 2) will be assumed per kg
	(dimension 2) of secodary production of product meat
	(dimension~3)

Further described in:

	Majeau-Bettez, G., R. Wood, E.G. Hertwich, and A.H. Strømman.
	2014. When do allocations and constructs respect material,
	energy, financial, and production balances in LCA and EEIO?
	Journal of Industrial Ecology. In Press: In Press.

Combines modifications from commits:
	031b9d4c7bade941016b1a7dbdfb72182d66ba80
	344338ec07a511f253cca48ead2a364b2bf08c6b
	0795546905fcdc14a03a93b14a185171f416c2cd
	in /K/indecol/Python/Modules/

 Changes to be committed:
	modified:   allocation_construct.py
	modified:   readMe_Variables.txt